+++ ChangeLog.md
@@ -52,6 +52,8 @@
 * Add `anyChildFocused`, which provides information about whether subwidgets are focused
 * Add `filterKeys`, which is the same as `localInput` but only cares about keyboard events
 * _Breaking Change_: `ImageWriter` now has a method `mapImages` for transforming the images emitted by a child widget
+* Add `hoistRunLayout` to apply a transformation to the context of a `Layout` action and run that action
+* Add an `MFunctor` instance for `VtyWidget`
 
 ## 0.1.4.1
 * Migrate to new dependent-sum / dependent-map (after the "some" package split)